Corrigido imcompatibilidade de argumentos da declaração das funções 'interfaceMenu' e 'interfaceString' e suas implementações

+++ include/interface.h
@@ -14,8 +14,8 @@
 
 void printCenterMessage(const char *msg, WINDOW *window);
 void handleWithSmallTermSize(int y, int x);
-int interfaceMenu(const char *title, const size_t n, ...);
-void interfaceString(const char *title, const size_t n, ...);
+int interfaceMenu(const char *title, const int n, ...);
+void interfaceString(const char *title, const int n, ...);
 void interfaceGetStringInput(char *msg, char *destiny, size_t lenght);
 void optionsMenu(void);
 void createProfile(void);
